requirements.txt
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
matplotlib>=3.7.0
pandas>=2.0.0
fake-useragent>=1.4.0
feedparser>=6.0.10
python-dateutil>=2.8.2
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
gunicorn>=21.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
redis>=5.0.0
flask-caching>=2.1.0
flask-jwt-extended>=4.5.0
flask-cors>=4.0.0
flask-swagger-ui>=4.11.0
flask-mail>=0.9.1
python-telegram-bot>=20.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
reportlab>=4.0.0
xgboost>=2.0.0
flask-limiter>=3.5.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
selenium>=4.15.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
vkontakte-api>=11.9.1

